+++ src/Domain/Entities/Comment.cs
@@ -4,10 +4,10 @@ namespace PersonalBlog.Domain.Entities;
 
 public class Comment : EntityBase
 {
-    public long PostId { get; set; }
+    public int PostId { get; set; }
     public Post Post { get; set; } = default!;
 
-    public long? ParentCommentId { get; set; }
+    public int? ParentCommentId { get; set; }
     public Comment? ParentComment { get; set; }
     public ICollection<Comment> Replies { get; set; } = new List<Comment>();
